Extract Config as Configuration to its own module.
This moves the types built up from it for use in planner sample inputs
off to their own targets module.
