+++ tests/python/pants_test/engine/exp/test_configuration.py
@@ -8,24 +8,24 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 import unittest
 
 from pants.base.address import Address
+from pants.engine.exp.configuration import Configuration
 from pants.engine.exp.objects import ValidationError
-from pants.engine.exp.targets import Config
 
 
-class ConfigTest(unittest.TestCase):
+class ConfigurationTest(unittest.TestCase):
   def test_address_no_name(self):
-    config = Config(address=Address.parse('a:b'))
+    config = Configuration(address=Address.parse('a:b'))
     self.assertEqual('b', config.name)
 
   def test_address_name_conflict(self):
     with self.assertRaises(ValidationError):
-      Config(name='a', address=Address.parse('a:b'))
+      Configuration(name='a', address=Address.parse('a:b'))
 
   def test_typename(self):
-    self.assertEqual('Config', Config().typename)
-    self.assertEqual('aliased', Config(typename='aliased').typename)
+    self.assertEqual('Configuration', Configuration().typename)
+    self.assertEqual('aliased', Configuration(typename='aliased').typename)
 
-    class Subclass(Config):
+    class Subclass(Configuration):
       pass
 
     self.assertEqual('Subclass', Subclass().typename)
@@ -34,35 +34,44 @@ class ConfigTest(unittest.TestCase):
   def test_extend_and_merge(self):
     # Resolution should be lazy, so - although its invalid to both extend and merge, we should be
     # able to create the config.
-    config = Config(extends=Config(), merges=Config())
+    config = Configuration(extends=Configuration(), merges=Configuration())
     with self.assertRaises(ValidationError):
       # But we should fail when we go to actually inherit.
       config.create()
 
   def test_extend(self):
-    extends = Config(age=32, label='green', items=[],
-                     extends=Config(age=42, other=True, items=[1, 2]))
+    extends = Configuration(age=32, label='green', items=[],
+                            extends=Configuration(age=42, other=True, items=[1, 2]))
 
     # Extension is lazy, so we don't pick up the other field yet.
-    self.assertNotEqual(Config(age=32, label='green', items=[], other=True), extends)
+    self.assertNotEqual(Configuration(age=32, label='green', items=[], other=True), extends)
 
     # But we do pick it up now.
-    self.assertEqual(Config(age=32, label='green', items=[], other=True), extends.create())
+    self.assertEqual(Configuration(age=32, label='green', items=[], other=True), extends.create())
 
   def test_merge(self):
-    merges = Config(age=32, items=[3], knobs={'b': False},
-                    merges=Config(age=42, other=True, items=[1, 2], knobs={'a': True, 'b': True}))
+    merges = Configuration(age=32, items=[3], knobs={'b': False},
+                           merges=Configuration(age=42,
+                                                other=True,
+                                                items=[1, 2],
+                                                knobs={'a': True, 'b': True}))
 
     # Merging is lazy, so we don't pick up the other field yet.
-    self.assertNotEqual(Config(age=32, items=[1, 2, 3], knobs={'a': True, 'b': False}, other=True),
+    self.assertNotEqual(Configuration(age=32,
+                                      items=[1, 2, 3],
+                                      knobs={'a': True, 'b': False},
+                                      other=True),
                         merges)
 
     # But we do pick it up now.
-    self.assertEqual(Config(age=32, items=[1, 2, 3], knobs={'a': True, 'b': False}, other=True),
+    self.assertEqual(Configuration(age=32,
+                                   items=[1, 2, 3],
+                                   knobs={'a': True, 'b': False},
+                                   other=True),
                      merges.create())
 
   def test_validate_concrete(self):
-    class Subclass(Config):
+    class Subclass(Configuration):
       def validate_concrete(self):
         if self.name != 'jake':
           self.report_validation_error('There is only one true good name.')
